separate vqbet util files
